add infrastructure_must_have_capabilities
